Initial Commit
An extension of a website I had to maintain for school, I am in the
process of extending it to be a permanent web presence for myself. The
original classwork will remain while I continue to add sections for my
own personal projects, which will vary from web-based to descriptions of
more traditional projects. It has been developed with ASP.NET and the
associated languages it needs to function. I will be adding new content
as I have the time and ideas.